Code Refactored for Guest service
