+++ guest-service/src/test/java/com/epam/guest/GuestApplicationTests.java
@@ -1,10 +1,10 @@
-package com.epam.guestservice;
+package com.epam.guest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class GuestserviceApplicationTests {
+class GuestApplicationTests {
 
 	@Test
 	void contextLoads() {
