+++ guest-service/src/test/java/com/epam/guest/controller/GuestControllerTest.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice.controller;
+package com.epam.guest.controller;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,19 +9,22 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
+import com.epam.guest.model.CreditCard;
+import com.epam.guest.model.StayHistory;
 import com.epam.guest.model.User;
 import com.epam.guest.service.GuestService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-class GuestServiceControllerTest {
+class GuestControllerTest {
 	@Autowired
 	private MockMvc mockMvc;
